Target extraction without running make
Running `make` to get all targets may be quite cumbersome
in both resource and number of yielded targets. Therefore,
an option to allow the user to choose a simpler form of
target extraction which only parses the Makefile as it is.

Directives such as "include" and other derived targets will
not be visisble with this option set (which may very well be
what we want).

Fixes #8